apps/gadget-comparator-using-gemini-url-context/requirements.txt
streamlit
google-genai